requirements.txt
openai>=1.0.0
jinja2>=3.0.0
python-dotenv>=1.0.0
